+++ lib/yt/playlist_item.rb
@@ -1,9 +1,9 @@
 module Yt
-  # Provides methods to interact with YouTube playlists.
-  # @see https://developers.google.com/youtube/v3/docs/playlists
-  class Playlist
-    # @param [Hash] options the options to initialize a Playlist.
-    # @option options [String] :id The unique ID of a YouTube playlist.
+  # Provides methods to interact with YouTube playlist items.
+  # @see https://developers.google.com/youtube/v3/docs/playlistItems
+  class PlaylistItem
+    # @param [Hash] options the options to initialize a PlaylistItem.
+    # @option options [String] :id The unique ID of a YouTube playlist item.
     def initialize(options = {})
       @id = options[:id]
       @data = HashWithIndifferentAccess.new
@@ -14,30 +14,30 @@ module Yt
 
   ### ID
 
-    # @return [String] the playlist’s ID.
+    # @return [String] the item’s ID.
     def id
       @id
     end
 
   ### SNIPPET
 
-    # @return [String] the playlist’s title.
+    # @return [String] the item’s title.
     def title
       snippet['title']
     end
 
-    # @return [String] the playlist’s description.
+    # @return [String] the item’s description.
     def description
       snippet['description']
     end
 
-    # @return [Time] the date and time that the playlist was created.
+    # @return [Time] the date and time that the item was added to the playlist.
     def published_at
       Time.parse snippet['publishedAt']
     end
 
-    # Returns the URL of the playlist’s thumbnail.
-    # @param [Symbol, String] size The size of the playlist’s thumbnail.
+    # Returns the URL of the item’s thumbnail.
+    # @param [Symbol, String] size The size of the item’s thumbnail.
     # @return [String] if +size+ is +:default+, the URL of a 120x90px image.
     # @return [String] if +size+ is +:medium+, the URL of a 320x180px image.
     # @return [String] if +size+ is +:high+, the URL of a 480x360px image.
@@ -48,54 +48,52 @@ module Yt
       snippet['thumbnails'].fetch(size.to_s, {})['url']
     end
 
-    # @return [String] the ID of the channel that published the playlist.
+    # @return [String] the ID of the channel that the playlist belongs to.
     def channel_id
       snippet['channelId']
     end
 
-    # @return [String] the title of the channel that published the playlist.
+    # @return [String] the title of the channel that the playlist belongs to.
     def channel_title
       snippet['channelTitle']
     end
 
-    # def tags # not yet implemented, not sure how to set to test
-    # def default_language # not yet implemented, not sure how to set to test
-    # def localized # not yet implemented
+    # @return [String] the ID of the playlist that the item belongs to.    
+    def playlist_id
+      snippet['playlistId']
+    end
+
+    # @return [Integer] the order in which the item appears in the playlist. 
+    #   The value uses a zero-based index so the first item has a position of 0.
+    def position
+      snippet['position']
+    end
+
+    # @return [String] the ID of the video that the item refers to.
+    def video_id
+      snippet['resourceId']['videoId']
+    end
 
   ### STATUS
 
-    # @return [String] the playlist’s privacy status. Valid values are:
+    # @return [String] the item’s privacy status. Valid values are:
     #   +"private"+, +"public"+, and +"unlisted"+.
     def privacy_status
       status['privacyStatus']
-    end
-
-  ### CONTENT DETAILS
-
-    # @return [<Integer>] the number of videos in the playlist.
-    def item_count
-      content_details['itemCount']
-    end
-
-  ### ASSOCIATIONS
-
-    # @return [Yt::Relation<Yt::PlaylistItem>] the items of the playlist.
-    def items
-      @items ||= Relation.new(PlaylistItem) {|options| items_response options}
     end
 
   ### OTHERS
 
     # Specifies which parts of the video to fetch when hitting the data API.
     # @param [Array<Symbol, String>] parts The parts to fetch. Valid values
-    #   are: +:snippet+, +:status+, and +:content_details+.
+    #   are: +:snippet+ and +:status+.
     # @return [Yt::Video] itself.
     def select(*parts)
       @selected_data_parts = parts
       self
     end
 
-    # @return [String] a representation of the Yt::Playlist instance.
+    # @return [String] a representation of the Yt::PlaylistItem instance.
     def inspect
       "#<#{self.class} @id=#{@id}>"
     end
@@ -110,10 +108,6 @@ module Yt
 
     def status
       data_part :status
-    end
-
-    def content_details
-      data_part :content_details
     end
 
     def data_part(part)
@@ -140,22 +134,6 @@ module Yt
       part = parts.join ','
       query = {key: Yt.configuration.api_key, id: @id, part: part}.to_param
 
-      Net::HTTP::Get.new("/youtube/v3/playlists?#{query}").tap do |request|
-        request.initialize_http_header 'Content-Type' => 'application/json'
-      end
-    end
-
-  ### ASSOCIATIONS (PLAYLIST ITEMS)
-
-    def items_response(options = {})
-      Net::HTTP.start 'www.googleapis.com', 443, use_ssl: true do |http|
-        http.request items_request(options)
-      end.tap{|response| response.body = JSON response.body}
-    end
-
-    def items_request(options = {})
-      part = options[:parts].join ','
-      query = {key: Yt.configuration.api_key, playlistId: id, part: part, maxResults: [options[:limit], 50].min, pageToken: options[:offset]}.to_param
       Net::HTTP::Get.new("/youtube/v3/playlistItems?#{query}").tap do |request|
         request.initialize_http_header 'Content-Type' => 'application/json'
       end